hw-10/requirements.txt
mysql-connector-python==8.0.23
Flask==2.2.5
Werkzeug==2.2.3
